+++ src/se/var.rs
@@ -1,5 +1,5 @@
 use crate::{
-    de::{INNER_VALUE, UNFLATTEN_PREFIX},
+    de::INNER_VALUE,
     errors::{serialize::DeError, Error},
     events::{BytesEnd, BytesStart, Event},
     se::key::QNameSerializer,
@@ -7,7 +7,6 @@ use crate::{
     writer::Writer,
 };
 use serde::ser::{self, Serialize};
-use serde::Serializer as _;
 use std::io::Write;
 
 /// An implementation of `SerializeMap` for serializing to XML.
@@ -130,24 +129,18 @@ where
     ) -> Result<(), DeError> {
         // TODO: Inherit indentation state from self.parent.writer
         let writer = Writer::new(&mut self.buffer);
-        if key.starts_with(UNFLATTEN_PREFIX) {
-            let key = &key[UNFLATTEN_PREFIX.len()..];
-            let mut serializer = Serializer::with_root(writer, Some(key));
-            serializer.serialize_newtype_struct(key, value)?;
-            self.children.append(&mut self.buffer);
-        } else {
-            let mut serializer = Serializer::with_root(writer, Some(key));
-            value.serialize(&mut serializer)?;
 
-            if !self.buffer.is_empty() {
-                if self.buffer[0] == b'<' || key == INNER_VALUE {
-                    // Drains buffer, moves it to children
-                    self.children.append(&mut self.buffer);
-                } else {
-                    self.attrs
-                        .push_attribute((key.as_bytes(), self.buffer.as_ref()));
-                    self.buffer.clear();
-                }
+        let mut serializer = Serializer::with_root(writer, Some(key));
+        value.serialize(&mut serializer)?;
+
+        if !self.buffer.is_empty() {
+            if self.buffer[0] == b'<' || key == INNER_VALUE {
+                // Drains buffer, moves it to children
+                self.children.append(&mut self.buffer);
+            } else {
+                self.attrs
+                    .push_attribute((key.as_bytes(), self.buffer.as_ref()));
+                self.buffer.clear();
             }
         }