+++ tests/test.rs
@@ -5,9 +5,6 @@ use quick_xml::reader::Reader;
 use quick_xml::Error;
 use std::borrow::Cow;
 
-#[cfg(feature = "serialize")]
-use serde::Deserialize;
-
 use pretty_assertions::assert_eq;
 
 #[test]
@@ -186,72 +183,5 @@ fn test_issue299() -> Result<(), Error> {
             _ => (),
         }
     }
-    Ok(())
-}
-
-#[cfg(feature = "serialize")]
-#[test]
-fn test_issue305_unflatten_namespace() -> Result<(), quick_xml::DeError> {
-    use quick_xml::de::from_str;
-
-    #[derive(Deserialize, Debug, PartialEq)]
-    struct NamespaceBug {
-        #[serde(rename = "$unflatten=d:test2")]
-        test2: String,
-    }
-
-    let namespace_bug: NamespaceBug = from_str(
-        r#"
-    <?xml version="1.0" encoding="UTF-8"?>
-    <d:test xmlns:d="works">
-        <d:test2>doesntwork</d:test2>
-    </d:test>"#,
-    )?;
-
-    assert_eq!(
-        namespace_bug,
-        NamespaceBug {
-            test2: "doesntwork".into(),
-        }
-    );
-
-    Ok(())
-}
-
-#[cfg(feature = "serialize")]
-#[test]
-fn test_issue305_unflatten_nesting() -> Result<(), quick_xml::DeError> {
-    use quick_xml::de::from_str;
-
-    #[derive(Deserialize, Debug, PartialEq)]
-    struct InnerNestingBug {}
-
-    #[derive(Deserialize, Debug, PartialEq)]
-    struct NestingBug {
-        // comment out one of these fields and it works
-        #[serde(rename = "$unflatten=outer1")]
-        outer1: InnerNestingBug,
-
-        #[serde(rename = "$unflatten=outer2")]
-        outer2: String,
-    }
-
-    let nesting_bug: NestingBug = from_str::<NestingBug>(
-        r#"
-    <?xml version="1.0" encoding="UTF-8"?>
-    <root>
-        <outer1></outer1>
-        <outer2></outer2>
-    </root>"#,
-    )?;
-
-    assert_eq!(
-        nesting_bug,
-        NestingBug {
-            outer1: InnerNestingBug {},
-            outer2: "".into(),
-        }
-    );
-
     Ok(())
 }